Fix issue with modulator when using const as order
When a constant is used in place of the modulator order, its value gets converted to a float
and then a string, which the modulator in Pykat tries to convert to int. Converting a string
'1.0' to int throws an error. This change parses the string as a float, then checks if it is
an integer, throws an error if not, then converts it to an int. It's hacky but it works.

**It's likely other code suffers from this bug if it copied the same parse code**